SetupTool: change index template class implementation
Here I've changed the implementation of the simple 3d index/vector template class to use an array of length 3 internally. This is mainly to ensure stricted adherence to the standard instead of relying on compiler specific behaviour (assuming that simple classes are laid out like POD). Use of access via ind.x etc no longer allowed, so the user of the class have also been updated, as has the python wrapper.